Optimised flatlists - fixed scrolling issue
