+++ app/components/ManageSpecies/index.js
@@ -1,15 +1,15 @@
 import { useNavigation } from '@react-navigation/native';
 import i18next from 'i18next';
 import React, { useContext, useEffect, useState } from 'react';
-import { Alert, FlatList, StyleSheet, Text, TextInput, TouchableOpacity, View } from 'react-native';
-import Icon from 'react-native-vector-icons/Feather';
+import { Alert,  StyleSheet,  TextInput,  View } from 'react-native';
 import Ionicons from 'react-native-vector-icons/Ionicons';
+import SearchSpecies from './SearchSpecies';
+import MySpecies from './MySpecies';
 import { Colors, Typography } from '_styles';
 import { addMultipleTreesSpecie, setSpecieId } from '../../actions/species';
 import { SpeciesContext } from '../../reducers/species';
 import { AddUserSpecies, getAllSpecies, searchSpecies } from '../../repositories/species';
 import { Header } from '../Common';
-import PrimaryButton from '../Common/PrimaryButton';
 
 const ManageSpecies = ({
   onPressSpeciesSingle,
@@ -32,54 +32,6 @@ const ManageSpecies = ({
 
   const onPressHome = () => {
     navigation.navigate('MainScreen');
-  };
-
-  const renderSpecieCard = ({ item, index }) => {
-    return (
-      <TouchableOpacity
-        key={index}
-        style={{
-          paddingVertical: 20,
-          paddingRight: 10,
-          borderBottomWidth: 1,
-          borderColor: '#E1E0E061',
-          flexDirection: 'row',
-          alignItems: 'center',
-          justifyContent: 'space-between',
-        }}
-        onPress={() => {
-          if (registrationType == 'single') {
-            onPressSpeciesSingle(item);
-          } else if (registrationType == 'multiple') {
-            onPressSpeciesMultiple(item, index);
-          }
-        }}>
-        <View>
-          <Text
-            style={{
-              fontSize: Typography.FONT_SIZE_16,
-              fontFamily: Typography.FONT_FAMILY_REGULAR,
-            }}>
-            {item.scientificName}
-          </Text>
-        </View>
-        {registrationType == 'multiple' ? (
-          <Text>{item.treeCount ? item.treeCount : 'NA'}</Text>
-        ) : (
-          <Ionicons name="chevron-forward-outline" size={20} />
-        )}
-      </TouchableOpacity>
-    );
-  };
-
-  const addSpecies = (item) => {
-    setSelectedSpecies([...selectedSpecies, item]);
-  };
-
-  const removeSpecies = (item) => {
-    setSelectedSpecies(
-      selectedSpecies.filter((specie) => specie.scientific_name !== item.scientific_name),
-    );
   };
 
   const addSelectedSpecies = () => {
@@ -121,130 +73,6 @@ const ManageSpecies = ({
     }
   };
 
-  const checkIsSpeciePresent = (speciesList, specieToSearch) => {
-    let isPresent = false;
-    if (speciesList && speciesList.length > 0) {
-      for (let specie of speciesList) {
-        if (specie.scientific_name === specieToSearch.scientific_name) {
-          isPresent = true;
-          break;
-        }
-      }
-    }
-    return isPresent;
-  };
-
-  const renderSearchSpecieCard = ({ item, index }) => {
-    let isDisabled = false;
-    let isCheck = checkIsSpeciePresent(selectedSpecies, item);
-    let isUserSpeciePresent = false;
-    if (specieList && specieList.length > 0) {
-      for (let specie of specieList) {
-        if (specie.speciesId === item.guid) {
-          isUserSpeciePresent = true;
-          break;
-        }
-      }
-    }
-    if (isUserSpeciePresent) {
-      isDisabled = true;
-      isCheck = true;
-    }
-
-    const SpecieListItem = () => {
-      return (
-        <>
-          <View>
-            <Text
-              style={{
-                fontSize: Typography.FONT_SIZE_16,
-                fontFamily: Typography.FONT_FAMILY_REGULAR,
-              }}>
-              {item.scientific_name}
-            </Text>
-          </View>
-          <Icon
-            name={isCheck ? 'check-circle' : 'plus-circle'}
-            size={25}
-            color={isCheck ? Colors.PRIMARY : Colors.TEXT_COLOR}
-          />
-        </>
-      );
-    };
-
-    if (isDisabled) {
-      return (
-        <View style={[styles.specieListItem, { opacity: 0.5 }]}>
-          <SpecieListItem />
-        </View>
-      );
-    } else {
-      return (
-        <TouchableOpacity
-          key={index}
-          style={styles.specieListItem}
-          onPress={() => {
-            if (isCheck) {
-              removeSpecies(item);
-            } else {
-              addSpecies(item);
-            }
-          }}>
-          <SpecieListItem />
-        </TouchableOpacity>
-      );
-    }
-  };
-
-  const MySpecies = () => {
-    return (
-      <View style={{ flex: 1 }}>
-        <View>
-          <Text
-            style={{
-              paddingTop: 25,
-              paddingBottom: 15,
-              fontFamily: Typography.FONT_FAMILY_BOLD,
-              fontSize: Typography.FONT_SIZE_16,
-            }}>
-            {i18next.t('label.select_species_my_species')}
-          </Text>
-        </View>
-        <View style={{ flex: 1 }}>
-          <FlatList
-            style={{ flex: 1 }}
-            data={registrationType === 'multiple' ? speciesState.multipleTreesSpecies : specieList}
-            showsVerticalScrollIndicator={false}
-            keyExtractor={(item) => item.speciesId}
-            renderItem={renderSpecieCard}
-          />
-        </View>
-        {registrationType === 'multiple' && (
-          <PrimaryButton
-            onPress={onSaveMultipleSpecies}
-            btnText={i18next.t('label.select_species_btn_text')}
-            testID={'btn_save_and_continue_species'}
-            accessibilityLabel={'Save and Continue Species'}
-          />
-        )}
-      </View>
-    );
-  };
-
-  const SearchSpecies = () => {
-    return (
-      <View style={{ flex: 1, paddingTop: 15 }}>
-        <FlatList
-          style={{ flex: 1 }}
-          data={searchList}
-          showsVerticalScrollIndicator={false}
-          keyExtractor={(item) => item.guid}
-          renderItem={renderSearchSpecieCard}
-        />
-      </View>
-    );
-  };
-
   const handleSpeciesSearch = (text) => {
     setSearchText(text);
     if (text) {
@@ -277,7 +105,9 @@ const ManageSpecies = ({
           onFocus={() => setSearchBarFocused(true)}
         />
       </View>
-      {searchBarFocused ? <SearchSpecies /> : <MySpecies />}
+      {searchBarFocused ? 
+        <SearchSpecies setSelectedSpecies={setSelectedSpecies} selectedSpecies={selectedSpecies} specieList={specieList} searchList={searchList} /> : 
+        <MySpecies onSaveMultipleSpecies={onSaveMultipleSpecies}  registrationType={registrationType}  speciesState={speciesState} onPressSpeciesSingle={onPressSpeciesSingle} onPressSpeciesMultiple={onPressSpeciesMultiple} specieList={specieList} />}
     </View>
   );
 };
